Release 1.11.1

+++ tests/t-fips-service-ind.c
@@ -1010,7 +1010,7 @@ check_cipher_o_s_e_d_c (int reject)
 
           if (taglen != tv[tvidx].taglen)
               fail ("gcry_cipher_info %d failed: taglen mismatch %d != %ld\n", tvidx,
-                    tv[tvidx].taglen, taglen);
+                    tv[tvidx].taglen, (long)taglen);
         }
 
       err = gcry_cipher_encrypt (h, out, MAX_DATA_LEN, pt, blklen);
